docs/tutorial-branches/chapter-00-introduction/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic_framework_pkg"
version = "0.1.0"
description = "A simplified agentic framework for querying CSVs with SQL."
authors = [
  { name = "Paul Rigor", email = "paul.rigor@pnnl.gov" },
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "fastmcp>=2.6.0",
    "mcp",
    "litellm>=1.15.0",
    "langchain>=0.2.43",
    "langchain-openai",
    "langchain-community",
    "langchain-google-vertexai",
    "langchain-ollama",
    "streamlit>=1.20.0",
    "aiosqlite>=0.17.0",
    "pandas>=1.5.0",
    "uvicorn[standard]>=0.20.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.20.0",
    "fastapi",
    "openai",
    "sqlalchemy",
    "chromadb",
    "langgraph",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "pytest-mock>=3.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "pip-audit>=2.5.0",
    "langchain-ollama",
]

[project.scripts]
run-mcp-server = "agentic_framework_pkg.mcp_server.main:start_server_cli"
run-streamlit-harness = "agentic_framework_pkg.streamlit_app.main:start_harness_cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E"]
ignore = ["E501"]
